le dashboard client + fontionnalite de reserve et charger contrat et paiement et charger la facture

+++ src/app/modules/admin/components/list-cars/list-cars.component.ts
@@ -2,17 +2,18 @@ import { Component ,OnInit } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { AdminService } from '../../services/admin.service';
 import { VoitureModele } from '../../modele/voiture.modele.model';
+import { AddCarComponent } from '../add-car/add-car.component';
 
 @Component({
   selector: 'app-list-cars',
   standalone: true,
-  imports: [CommonModule],
+  imports: [CommonModule,AddCarComponent ],
   templateUrl: './list-cars.component.html',
   styleUrl: './list-cars.component.scss'
 })
 export class ListCarsComponent implements OnInit {
   voitures: VoitureModele[] = []; // Liste des voitures
-
+  showAddCarModal: boolean = false; //
   constructor(private voitureService: AdminService) {}
 
   // Initialisation : Charger les voitures depuis le backend
@@ -32,7 +33,7 @@ export class ListCarsComponent implements OnInit {
     );
   }
 
-  // Ouvre le formulaire d'ajout de voiture
+ /* // Ouvre le formulaire d'ajout de voiture
   openAddCarForm(): void {
     console.log('Formulaire d’ajout de voiture ouvert');
     // Logique supplémentaire à implémenter (ex : afficher un modal)
@@ -42,12 +43,24 @@ export class ListCarsComponent implements OnInit {
   editCar(voiture: VoitureModele): void {
     console.log('Modification de la voiture:', voiture);
     // Logique supplémentaire à implémenter (ex : remplir un formulaire)
-  }
+  }*/
+    openModal(): void {
+      this.showAddCarModal = true;
+    }
+  
+    closeModal(): void {
+      this.showAddCarModal = false;
+    }
 
-  // Supprimer une voiture par son ID
-  deleteCar(id: number): void {
+  // Supprimer une voiture par son ID avec confirmation
+deleteCar(id: number): void {
+  // Affichage du message de confirmation
+  const confirmation = confirm('Êtes-vous sûr de vouloir supprimer cette voiture ?');
+
+  if (confirmation) {
     this.voitureService.deleteVoiture(id).subscribe(
       () => {
+        // Mise à jour de la liste des voitures après suppression
         this.voitures = this.voitures.filter(voiture => voiture.id !== id);
         console.log('Voiture supprimée avec succès');
       },
@@ -55,6 +68,10 @@ export class ListCarsComponent implements OnInit {
         console.error('Erreur lors de la suppression de la voiture:', error);
       }
     );
+  } else {
+    console.log('Suppression annulée');
   }
+}
+
 
 }
